refactor(observability): derive capture directives, split the filter module
Addresses both findings from the review on #588.

**1. The ERROR branch hardcoded the capture crate names**, so a target added
to the list - which this PR's own CLAUDE.md guidance instructs - would have
had coverage at DEBUG/INFO/WARN and silently none at ERROR. The reviewer also
noted the test would not have caught it, since it asserted on the two literal
crate names.

Fixed at the root rather than by patching the branch: `CAPTURE_DIRECTIVES`
(a pre-formatted string) becomes `CAPTURE_TARGETS` (bare crate names), and
each branch supplies only the level. There is now no place where a target and
a level are written together, so the drift is not expressible.

The test now iterates `CAPTURE_TARGETS` for directive presence at every level.
Verified it actually fails on the described scenario by reintroducing the
hardcoded branch plus a third target:

    ERROR filter has no directive for screenpipe_capture:
      meridian=error,sqlx=error,screenpipe_screen=error,screenpipe_a11y=error

The behavioural half still names two targets because `tracing::error!`
requires a literal target; a comment records that the loop is the part which
scales, so the next reader does not mistake the literals for the guard.

**2. `src/observability/mod.rs` had grown to 622 lines**, over the repo's
500-line cap. Extracted the filter into `observability/filter.rs` - the
`EnvFilter` construction, the reload handle, and their tests - following the
same split as `install_mode.rs` and `otlp_target.rs`. "What may be captured"
is self-contained and independent of the subscriber/exporter wiring.

    mod.rs      622 -> 447   (below its 465 pre-PR size)
    filter.rs         244
    otlp_target.rs    276
    install_mode.rs    59

Verified: daemon fmt, clippy -D warnings, 774 tests.
